[server] move files
